requirements.txt
adal==1.2.5
aiohttp==3.7.3
alembic==1.4.3
ansiwrap==0.8.4
apache-airflow==1.10.12
apispec==1.3.3
appdirs==1.4.4
appnope==0.1.0
argcomplete==1.12.2
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
azure-core==1.9.0
azure-datalake-store==0.0.51
azure-storage-blob==12.6.0
Babel==2.9.0
backcall==0.2.0
black==20.8b1
boto3==1.16.25
botocore==1.19.25
cached-property==1.5.2
cachetools==4.1.1
cattrs==1.0.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.36
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
dnspython==2.0.0
docutils==0.16
email-validator==1.1.2
entrypoints==0.3
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
fsspec==0.8.4
funcsigs==1.0.2
future==0.18.2
gcsfs==0.7.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
graphviz==0.15
gunicorn==20.0.4
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
iso8601==0.1.13
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
lazy-object-proxy==1.5.2
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
msrest==0.6.19
multidict==5.0.2
mypy-extensions==0.4.3
natsort==7.1.0
nbclient==0.5.1
nbformat==5.0.8
nest-asyncio==1.4.3
nteract-scrapbook==0.4.1
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
papermill==2.2.2
parso==0.7.1
pathspec==0.8.1
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
prison==0.1.3
prompt-toolkit==3.0.8
psutil==5.7.3
ptyprocess==0.6.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
setproctitle==1.2
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
tabulate==0.8.7
tenacity==4.12.0
text-unidecode==1.3
textwrap3==0.9.2
thrift==0.13.0
toml==0.10.2
tornado==6.1
tqdm==4.54.0
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==0.16.1
WTForms==2.3.3
yarl==1.6.3
zope.deprecation==4.4.0
